Provide a container for client static files
In order to ensure that the environment running application code is
stable and portable, a Docker container is provided.

In the future, the static files will most likely be copied from the
output of a build process rather than a source directory, which is being
used right now to simplify the initial deployment.