Provide manifest file for droplet deployments
In order to leverage Terraform for the deployment of droplets, a
manifest file has been provided with a basic configuration for the
deployment of a small droplet in the NYC region.

This will need to be followed up with the provisioning of Docker on said
deployment so as to execute the release image to finalize the deployment
process. This may replace DigitalOcean with AWS for subsequent
deployments.