Provide packages & development scripts for building React app
The included packages were chosen to build a React application with WASM
support along with including a full test suite for verifying code.

Other npm scripts were added to start the development processes, as well
as some basic git hooks for now which may or may not be deleted later
depending on what the team ultimately decides on.