Define test coverage report rules
In order to ensure that a certain percentage level of test coverage is
met, a config file for the nyc library has been provided.

For now, there is no expected % of test coverage to be passed in order
to have successful code pushes, but that should change in the future
depending on what the team decides on.