Define a webpack config for development & production modes
In order to facilitate the development process and provide a means for
distributing source code in a production-ready manner, a webpack config
has been defined which allows the use of SASS/WASM/JSX.

At the moment there's an issue with the compilation phase for production
builds in where the WASM modules aren't properly being included. The
decision to fix this has been deferred for now since more progress may
be made with WASM to where this is finally supported on the client-side
of things and currently, it works in with the development build.